requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
plotly>=5.18.0
dash>=2.14.0
pandas>=2.1.0
numpy>=1.24.0
click>=8.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

